Data Science/Machine Learning/Student_mark_prediction/requirement.txt
panda
numpy
seaborn
matplotlib
scikit-learn
#-e .